requirements.txt
Pillow
pyheif
piexif
tqdm